Resolve regression in test_recursive_leader1.
In the canInlineIterator() predicate, the Vec being iterated over has new items appended
during iteration.  The iterator for std::vector does not support this usage, which leads
to memory corruption.  To get around this, I rewrote the current control structure, which
is effectively the inline unrolloing of a recursive call as (you guessed it!) a recursive
call.