Allow LICM to hoist some global variables
This is a conservative modification that will allow LICM to hoist invariant
globals from loops that do not contain any function calls.

Currently we mark that all global variables might have be modified elsewhere
(outside the immediate scope of the loop that we're hoisting from.) Because of
this no global variables were considered as candidates for LICM. There are
three ways a global can be modified:
 - Directly in the loop
 - Indirectly in the loop by a possibly nested function call
 - By some other task anywhere in the program

We already detect if the current loop directly modifies a global. If some other
task modifies a global, then in order for the program to be correct a
sync/single or atomic must be used. Currently LICM will not be performed on
functions that contain any sync/single or atomic variables so we don't have to
worry about this case. The currently un-handled case is if a possibly nested
function call inside of the loop modifies the global.

This patch updates marking that globals might have been defined elsewhere to
only do such marking if there are any functions calls inside of the loop. We
could be more aggressive and do some inter-procedural analysis to find out if
the function call actually modified the global but this will take care of most
cases, and is a much cheaper computation. It is also much less of a code
change, which is nice this close to release.

This is primarily for tests that do not define a main procedure and instead use
the implicit main. This change is motivated by a performance regression with
some of our slower fannkuck versions from the last update to LICM
(75ada8e3483f)