Add a new test for the root case.
The /dev case doesn't apply to cygwin, but it seems useful to check both the
root and another known mount point (because they are handled distinctly), so
instead of making work for myself later, I just copied the test over and
replaced "/dev" with just "/".  Now we can check that cygwin detects something
right, and I'll expand testing further in the future.