Update range follower so zippered range iterators can be inlined
We saw a huge performance difference between parallel iteration over a range vs
domain that was resolved with 023b1bd838a3d64289a1c6292dce4a1df2e57038. We
noticed there was a similar performance regression for zippered range vs domain
parallel iteration. The root problem in both cases is in the same. Constructs
in the follower iterator prevented it from being inlined. In the non-zippered
case there was a second yield which prevented inlining. In the zippered case
the early returns for 'flwlen == 0' prevent inlining which kills performance.

This updates the range follower to to allow zippered iterator inlining to fire,
which get parallel zippered range iteration on par with domain iteration.

In order to do this all of the early returns are removed. The early returns
were all for if the followers length was 0. What we do instead is for strided
and non-stride cases we create a 0 length range to iterate over by default, and
we only replace that with the real range if the followers length was not 0.
It's not a pretty solution, but it's the cleanest solution I could come up with
with the current zippered iterator inlining requirements. This resolves the
performance gap for test/performance/bharshbarg/*