Pipe new docs args through main function.
Add new helper function to arg.h/cpp, init_arg_desc() that takes an
ArgumentState struct and sets its desc member with the given
ArgumentDescription.

Update fDocs to be set solely off the program name. This is the only way
it is set now.

Update main() to call init_arg_desc() with docs_arg_desc when fDocs is
true. Otherwise, call it with arg_desc as before. This means the arguments
defined in docsDriver.cpp will be used when argv[0] is chpldoc. Otherwise, the
regular arguments defined in driver.cpp will be used.

Create new fDocsPrintHelper global that is set when chpldoc args include
-h,--help flag. Pipe flag through to the usage() call.