Qthreads: don't use guard pages when the needed mprotect(2) won't work.
Qthreads uses unreferenceable guard pages at the ends of qthread (Chapel
task) stacks to provide stack overflow protection.  But the mprotect(2)
sys call to make these guard pages unreferenceable is invalid when the
memory involved comes from hugepage space.  Qthreads continues execution
when those mprotect(2) calls fail, but it also produces error messages
about the failures.  For various reasons we often use hugepages on Cray
X* systems.  To avoid the noisy output, here we turn off guard pages in
Qthreads whenever the heap page size appears not to be the same as the
system page size.

There are also supporting changes in chplsys.{h,c}, to return the system
and heap page sizes.  Note that the logic for the latter is not totally
solid at this point and we may need to revisit it at some point.  For
example, it doesn't catch the case in which we have a hugepage module
loaded but have not enabled automatic placement of the program's heap on
hugepages.  However, certainly for now this will do.

There is future follow-up work here.  For various reasons, in several
places in the runtime we acquire the page size.  These should be changed
to call the appropriate new chplsys.c function.  However, that work can
wait until after the upcoming release is done.