Do not end enclosing scopes in the repeated clauses of loops.
When associating scopes with basic blocks, do not mark enclosing scopes if the last
statement in a give block is contained in one of the repeated sections of a loop.  This
avoids calling autoDestroys repeatedly (within the loop) on variables that are declared
outside the loop.

Down to 251 regressions (80 run-time).  I'm counting this as a win even though the number
of runtime regressions went up.  (Further analysis required.)