Implement curDir and parentDir
These simple functions return the token representing abstractly the current
directory and the parent directory, or `.` and `..` respectively on Windows and
POSIX systems.  I didn't implement the portion for other systems, since the
Wikipedia article I found listing other values were made up of deprecated
systems and those I don't think Chapel itself is likely to build correctly on
at the moment.  Should a new user base arise, it is hoped that someone familiar
with these systems would be able to add support for them and contribute it
back to use or at least give us pointers for how to detect that system.

I may add one or two other tests of this functionality.