Fix syntax error in start_test related to dir skipifs, like test/chpldoc.skipif.
Previously, if a dir skipif was executable and printed True or False (instead
of 1 or 0 respectively), the while loop broke because `while ( ! False && ...)`
is not valid csh. Similary, `while ( ! True && ...)` is not valid.

Update the body of while loop that calls the directory skipif to convert
True/False values to 1/0 values.