Update docs() pass to set PYTHONPATH before calling sphinx-build.
In our linux testing environment, the sphinx-build program fails because it
cannot find the sphinx module. This happens because PYTHONPATH is set to
something else and not overridden/added to.

Update the docs() pass to set PYTHONPATH to the lib/python*/site-packages dir
in the venvDir. This ensures sphinx-build will search for sphinx module in
correct location.