Fix coverity, llvm build errors, and sub_test warning message.
Remove the getPythonPath() function from docs.cpp and the associated code for
setting PYTHONPATH.

Previously, I thought this fixed the build on our test systems. It turns out
that was not the case, as the python path settings were very strange on that
system.

In general, setting the PYTHONPATH should not be necessary as long as the PATH
is set and the correct sphinx-build is used. We might need to ensure PYTHONPATH
is unset, but that is not clear yet.

Also, fix sub_test warning to include line break and information about test
that is being run.