Fix build problems with LLVM+RE2 on Mac OS X
After PR #1539, we build third-party libraries (including RE2)
with the included clang compiler to support --llvm. When we
do that on Mac OS X, we set CXX (and CC) to be more than
one word, which breaks because we weren't properly quoting
those variables in third-party/re2/Makefile. Here we add
single quotes (as that is how the other third-party Makefiles
do it) and change the other quoted variables to use single
quotes as well.