Comment out unused variables used in a comment, add clang-included
I checked for other cases in util/chplenv where we do something
special for clang, and this came up in my search. So I added
clang-included there too, but then realized that none of that
code is used right now (it is used only in a commented out block).
So I commented out that code as well; when the block below
is uncommented these variable settings should be too.