Change the correct line in the makefile...
Make the LLVM runtime build optional based on feedback from Thomas

If you try to chpl -llvm and the runtime with clang-included is
not built, you do get a decent error from the compiler. But
some LLVM-based features will still work (extern blocks) if
the clang-included runtime does not build.