Fix types/string/bwross/string_index.chpl.
Passing needToCopy=false to reinitString needs to be used with extreme caution.  It can
only be used in the case where the buffer being passed in is owned in the calling
context.  Which means in particular, that passing in the string.buf field from another
string and setting needToCopy = false will definitely result in read-after-free errors.

I scanned and fixed all of the cases that were obviously in error.  The problem should
only arise again w.r.t. new uses of reinitString.