Improvements to sameFile (trivial, appearance related)
Michael updated the file utilities to use 'out error: syserr' instead of 'out
err: syserr' for consistency with the IO module.  Since the documentation
comments are my area for the release, I'm updating it to match the new argument
name.

I also improved the error message generated to be clearer.