Fix an overlapping memcpy
When I was doing some debugging with the madness test, I
noticed that the #line numbers generated were wrong.

To reproduce:
  cd test/studies/madness/common/
  chpl test_mul.chpl --savec=tmp --cpp-lines
  view tmp/MRA.chpl - are the #lines using MRA.chpl or MRA.clpl?

I checked that build with:
  valgrind chpl test_mul.chpl  --savec=tmp --cpp-lines

which reported
  ... Source and destination overlap in strcpy ... beautify.cpp:218
So, I have replaced that strcpy with an equivalent memmove (when the
strings do not overlap). I verified that fixed the valgrind error and
that the #lines for MRA.chpl are no longer garbled.