Speculatively build gmp and re2 with clang-include compiler.
Previously, gmp and re2 could only be built manually with the clang-included
target compiler. Update the llvm-runtime-if-needed target name to call the
third-party-try-opt target, which speculatively builds gmp and re2 (and
possibly others in the future).

This makes re2 and gmp functionality available to llvm. gmp does not yet work
with llvm, but it hopefully will some day!