Discovered problem with file portion of copyTree and fixed the test case.
Was failing to add the src path to the filename to copy.  Fixed now.

Test case takes a directory that has three layers, each with a file in it, and
copies that over.  I have tested the case where something that deep did not have
files in it.  Will test further, but I am fairly confident it works now.