Allow CHPL_TEST_NUM_TRIALS to be set outside of start_test
Traditionally CHPL_TEST_NUM_TRIALS could only be set by start_test and was just
used to communicate the --numtrials start_test option to sub_test. This allows
the number of trials to be set by the "user" with the env var and start_test
won't override it.

This is really just for xc performance testing so we can do multiple trials
form a place where we don't have an easy way to add additional start_test args