Parallel walk() improvements: parallel at all levels, only one sort warning
It turns out that the standalone walk iterator can be recursive as
long as it's called explicitly in a 'for' loop passing along the
tag rather than through a forall.  I could've sworn I tried this
the other night, but apparently not.

It also turns out that I had two warnings about sorted output, only
one of which was guarded against redundant calls, and not very well
(the boolean seemed backwards and it only worked because the call
itself was not recursive).  Here, I came up with a cleaner way of
dealing with that by simply passing 'sort=false' to the recursive
calls to the iterator.