Fix pining_for_lisp test cases using the "don't declare return value" workaround.
There is a logic error in normalize.cpp that affects return value type inference and
memory management.  Due to this interdependency, it remains unaddressed.  However, as part
of regularizing the constructor story in must eventually be done right.

The problem is that it injects the declared type of the return value by creating a return
value variable with the declared type.  Type resolution then requires that all return
value expressions be coercible to this type.  So far, so good.

At present, however, the return value variable is updated through assignment.  AMM demands
that the left operand of assignment be initialized.  The code that normalize emits for the
return value declaration omits the (default) initialization for the return value variable.

As a first step, we can add that initialization.  As a next step, we can remove that and
instead ensure that the return value variable is initialized (not assigned!) in the return
statement.