Fix chpldoc install for long directory paths.
Previously, when chapel was in a deep directory path, the pip install command
was failing because the hash bang exceeded the OS length and was truncated. For
example, in nightly testing, the pip install call was failing with:

  /bin/sh: .../path/to/bin/pip: <truncated hash bang>: bad interpreter: No such file or directory.

Update the chpldoc-venv/Makefile to invoke pip by calling python, avoiding the
hash bang altogether. The call is now `python .../path/to/bin/pip ...`.