Remove ../ from virtualenv dir paths.
The virtualenv paths are not normalized, so something with the form
a/b/../../a/b/ is added as hash bang to installed scripts. Start with the
normalized path in Makefile.include to avoid these issues altogether.

Specifically, the virtualenv --relocable call normalizes the path that it gets,
but then uses string comparisons when looking at hash bangs. So, it ends up not
normalizing some things, most importantly sphinx-build. By using the normalized
path from the beginning of time, this issue is avoid entirely.