Build clang-included runtime, et al with associated top-level targets.
Previously, the clang-included runtime and speculative third party builds were
a separate target, llvm-runtime-if-needed, which was invoked from comprt. This
worked well when the default target or the comprt target was called. However,
it was less than ideal when runtime or third-party-try-opt was invoked
directly.

For example, if a developer was working on the runtime and iteratively calling
`make runtime` from the top level, the clang-included runtime was not rebuilt
with each rebuild.

Update the makefile to build the clang-included runtime as a part of each of
the runtime and third-party-try-{gmp,re2} targets. Now, when the runtime is
built, the clang-included runtime will be built. The same is true for the
speculative gmp and re2 builds.

This will fix the issue with llvm and gasnet.llvm nightly testing configs.

* [ ] Call `make runtime` and ensure gen/ directories have regular and clang-included variations.
* [ ] Call `make third-party-try-opt` and ensure gmp/install/ and re2/install/ dirs have clang-included variations.