Report error and stop build when easy_install is not found.
Previously, there were make targets in third-party/chpldoc-venv/ that attempted
to check that python and easy_install were available. However, they did not
work in the case that $(which easy_install) returned nothing, since the
$(EASY_INSTALL) variable was empty, which meant $(EASY_INSTALL) as a dependency
did not mean anything.

Update third-party/chpldoc-venv/Makefile to use phony targets when run the
python and easy_install checks. This way, they will always work especially when
easy_install or python is not found.