Support --cflags in compileline with support for specialize, etc
Here I have enabled both runtime/etc/Makefile.include and
util/config/compileline to compute the equivalent to COMP_GEN_CFLAGS,
which I have named MAKE_COMP_GEN_CFLAGS. This is an alternative
to the way it is computed in the makefile generated in
files.cpp:codegen_makefile where COMP_GEN_CFLAGS is filled in
by the compiler in C++ code. Since the LLVM backend does not
use codegen_makefile however, it needs a different approach.
Previously it hardcoded values for clang, but that is not
an appropriate solution for SPECIALIZE_CFLAGS since they are
generated by a combination of Make scripts and Python.
Therefore, I have enabled compileline to produce this information
with the --cflags argument. The environment variable-setting
syntax for compileline provides the mechanism to tell the
Makefile which settings to use, as in
 compileline COMP_GEN_OPT=1 --cflags
will output the C compiler flags with optimization turned on.

I think it would be nice in the future to rely on
runtime/etc/Makefile.include to compute COMP_GEN_CFLAGS instead
of doing it in files.cpp. To do so we would need to set the
variables like COMP_GEN_OPT in files.cpp instead of computing
COMP_GEN_CFLAGS directly.