Add three new futures for chpldoc.
Mike discovered that default value output for reals, imags, and complexes were
handled incorrectly.  Since a good solution would involve more effort and
planning, we have decided to delay and implement it after the release, rather
than implement a poor solution that relies on behavior that will change in the
near future.  To make sure this is not forgotten, add a future to handle these
cases.