Alter failure detection for copyTree test.
Before, I was using the .catfiles to grab the contents of every text file in
the hierarchy that was to have been copied.  When copyTree was originally
reviewed, Vass pointed out that it would be more understandable and catch more
cases if there was a prediff that "diff"ed the source and destination directory.
So I removed the .catfiles, added a prediff, and updated the expected output.