This commit starts the effort to make the initial test more robust, and clearer.
As a consequence, a/ has been renamed to complexSrc, and complexSrc now has some
different children in it.  I will be removing all the single letter names, and
expanding the original test further (some action has been done already, but not
all of what was added.  I want to get the module documentation in today).