Add chpldoc documentation for Math.chpl.
The module was wrapped in a named module scope, to provide chpldoc with a place to hang
module-wide documentation.  The content was indented to reflect the added nesting level.

Entries were segregated into public and private interfaces.

Public interface entries were alphabetized using case-insensitive sorting.

Where the return value of a function could be determined uniquely, it was added to the
signature as an explicit return type.  Otherwise, the relation between the argument
type(s) and the result type is specified in the documentation.

Overall, a minimalist approach was used: If the reader can obtain the pertinent
information from the signature itself (e.g. the name and type of the arguments and the
type of the return value), these elements are not mentioned in the accompanying
documentation string.  This keeps the rendered documentation as concise as possible.