Simplified logic when 0 matches are found
On Elliot's suggestion, simplified iterator logic when glob() finds
zero matches in order to remove a conditional and generate cleaner
loops.  This requires doing some safe casts in order to avoid
subtracting -1 from a uint with possible value 0; and then to cast
the resulting int back to ssize_t.

Extended the testemptyglob test to support empty standalone, leader,
and follower cases, including a follower where the glob() is not
large enough.