Fix bug in localizeIteratorReturnSymbols().
In iterators where the return value type is declared, there will be
an initializer (constructor) call that initializes the return-value variable.
Corresponding updates of the RVV use assignment (rather than a simple MOVE primitive).
Therefore, yields and returns rely on the variable having been initialized.

localizeIteratorReturnSymbols() breaks this assumption because the
initialization is not copied when the local return symbol is created.   As
a result, we have to then look for places where the RVV is updated by
assignment and the original RVV is replaced by a new one.  We then bridge
out the assignment and insert a MOVE instead.

Note that the assignment is still required in the implementation of return
value inference.  It is inserted in normalize.cpp, but is no longer needed
after the resolution pass.

This change parallels one added recently #1600 for task intents on the mainline.

###DETAILS

Added comments to the code in normalize.cpp that causes a return-value variable to be
initialized at the top of a function that explicitly specifies the return type.

Added a clause in localizeReturnSymbols() to convert assignments to the RVV into a
PRIM_MOVE instead.  This is because assignment expects the LHS to be initialized, while
inserted "newRet" symbols are not initialized.  This (obviously) expects the value of the
RHS of the assignment to contain a fully-constructed object (in case it is of record
type).  The current implementation of AMM will insert a copy-constructor call, if needed,
to make this true.