Added an in-place version of plan_dft(); added a size-mismatch test
As discussed with Nikhil and Mike, it seemed asymmetric to not
support separate overloads for in- and out-of-place plan_dft()
even though it's not as necessary as in the r2c/c2r cases, so
this commit adds an overload.  Both overloads call a helper
routine to do the actual C plan_dft() call, in order to avoid
the size check in the in-place case.

It also adds a new test that checks for the size mismatch condition.