Fix isMount in the case of symlinks and test it
Previously, if a symlink was formed to a mountpoint, isMount would think that
symlink was also a mountpoint.  This is not correct behavior, with the rationale
being that if the original file was removed, the symlink would not maintain
the mount point.  Symlinks should have their link node compared instead.  My
method of finding the parent of a file also did not support this, as attaching
"/.." to the end of a symlink resulted in the parent of the location being
linked to instead of the parent of the location of the symlink itself.  Getting
around this required some string operations in the runtime which could be
replaced by functions in our Path module when they are implemented.