Support valgrind testing on chpldoc tests (take 2).
Previously, the chpldoc tests (i.e. those with .doc.chpl suffix) were using the
chpl compiler when running under valgrind due to an incorrect if/else block.
This led to invalid option errors, because chpl was called with chpldoc options
(e.g. --text-only).

Also (previously), sub_test was only using chpldoc when "compiling" the test, but
it was still passing chpl as the compiler to all the scripts.

Update sub_test to correctly wrap chpldoc compiler commands with valgrind and
set compiler variable for the rest of the test calls (e.g. to .prediff
etc). This requires compiler variable be reset at the beginning of each new
test to the original value. Add original_compiler variable that is set outside
the for loop and used to reset compiler with each iteration.

* [ ] full test pass
* [ ] Run `start_test -valgrind test/release/examples/primers/chpldoc.doc.chpl`
  and confirm it passes.