Bug fix for limiting the max size of qthreads memory pools
In cee911bf5a2e we added code to limit the max size of qthreads memory pools.
However, this code only limited the pool size when qthreads was adjusting the
initial alloc_size up if it thought it wasn't allocating enough bundles for
pooling to be worthwhile. It did not handle if the initial alloc_size was too
large.

Now if the initial alloc_size is too large, we round item_size up so it's a
multiple of pagesize (since qthreads expects allocations to be a multiple of
pagesize) and we find the max number items of that size that will fit within
max_alloc_size.

This bug was discovered while trying to figure out why ugni+qthreads seemed to
use so much more memory. The reason we noticed this bug there is because we ran
on some resource limited machines and guard pages don't work when ugni is being
used as the comm layer. In reality this was broken anytime we turned off guard
pages we just didn't notice it before.

With guard pages on, item_size just happened to work out to be a multiple of
pagesize so we got lucky and successfully limited the pool size. With guard
pages off stack size wasn't a multiple of pagesize, and the initial alloc_size
was over 1GB.