Small doc changes to the FileSystem and Path documentation.
@vasslitvinov pointed out that my use of code blocks for argument names was
inconsistent with the documentation recommendations.  While fixing that, I also
updated the umask comment so that it included the same warning chdir and cwd
generate and fixed the broken references to it (now that it is an operation on
a locale instead of just a plain function, it needs to include "locale." in
its reference).