Add sentence to reStructuredText primer about context.
The reStructuredText primer is largely based on the Sphinx and Python reST
primers, which also share a lot of information and text. The primer in
README.chpldoc.rst is catered to chpldoc use cases and therefore has a lot less
information. Some of the text in README.chpldoc.rst is the same as the Sphinx
and Python primers.