Couple of edits to the third-party README.
Update third-party README brief entry for chpldoc-venv to mention that the
packages are also downloaded and built when 'make chpldoc' is run.  Also,
updated the pointer for GMP's license files to reference COPYING.LESSERv3
instead of COPYING.LIB.  At some point between GMP 5 and 6 this was moved, but
we missed that and so were referencing a file that did not exist.