Fix intel performance failure with original chameneos
During performance runs with the intel compiler, we discovered that the original
chameneos shootout implementation was failing to get the correct number of
meetings.  In investigating, this seemed to be triggered by using casts and
arithmetic on the Color enum to determine the complement, instead of using a
select/if statement as required by the benchmark.  Because we had updated all
the other implementations to use the select statement, none of them showed
this problem.

It's possible that messing up the color to store caused overlap to the meetings
for the individual chameneos, as the two fields are right next to each other,
though that would depend on the endianness of the system for why it was only a
small number of mistakes instead of orders of magnitude off . . .

The right solution for the benchmark itself is to update to use the select
statement.  It may be worth investigating if something strange and mysterious is
happening to cause enums values to appear that shouldn't by the laws of
reasonable arithmetic . . .