Stop attaching the ivdep pragma to loops for CCE
We attach the ivdep pragma to loops that we know are order independent in order
to help the backend compiler vectorize. Unfortunately the Cray compiler has
slightly different semantics than we thought, so we can't throw it for the time
being. Each compiler vendor has subtly different semantics for the pragma.
Intel does not ignore proven vector dependencies, GCC only cares about
dependencies between loop iterations but not about dependencies for a single
iteration. Cray will ignore proven dependencies and cares about them for a
single iteration of the loop.

We had a regression with "release/examples/benchmarks/hpcc/fft" for single
locale performance testing that exposed this issue. The core issue here is
that while a loop was order independent it couldn't be legally vectorized.
However, the Cray compiler assumes the user meant what they said (it ignores
proven dependencies) and vectorizes the loop producing incorrect results. We
shouldn't be attaching the pragma to the loop in question, but that analysis to
determine that is far beyond what the current approach does.

The other compilers should be safe because of the different semantics mentioned
above.

See the Cray bugzilla entry for 824176 for more details.