Update what configurations use "fifo" as default tasking layer
Phew, we list when fifo is the default tasking layer in a bunch of places!

Previously it was just cygwin and knc. This release we made it the default when
cray is the target compiler (waiting on 8.4 to be released with support for gnu
style inline assembly.)

We also default to it for netbsd based on our limited experience building on
some netbsd boxes. I think qthreads will actually build on netbsd with a newer
gcc, but those systems don't have it and this is the solution until we build
qthreads speculatively or something.