Update hwloc notes in README.chplenv
Remove note saying that hwloc doesn't build on mac.

Add first draft of some new wording to indicate when hwloc is actually needed
(only for qthreads) and that if it doesn't build cleanly you should still be
able to run if you manually set it to "none" and rebuild, at least when you're
using flat. It *might* work for numa as well, but qthreads is really hampered
in what information it can get so it doesn't seem worthwhile to recommend.