Update chpl_launchcmd.py to work from nfs mounted lustre FS.
Previously, some nfs configuration issues were causing lots of failures with
XE testing when the launcher tried to access the executable. The specific
errors were "Text file busy", and this seemed to indicate that the file had
not been synchronized back to the lustre file system. Essentially, the file
was created successfully on the system with lustre mounted over nfs, but it
had not propagated back to the original lustre FS (as least not fully) before
qsub was trying to run ./execname.

It turns out, adding a simple "test -f ./execname" call before trying to run
./execname "fixes" or "works around" the issue. Update chpl_launchcmd.py to
include this check for all launchers.