Update sphinx project to support CHPLDOC_AUTHOR env var as author.
Previously, the literal string "AUTHOR TEXT" was used as the author in the HTML
(et al) output. There is work to make this more accessible via command line
flag and env var, but that will be committed for the next release. It seems
necessary to provide some way to change this value for the current, 1.11,
release.

Update the sphinx project conf.py to look for the author string in the
CHPLDOC_AUTHOR environment variable. If it is not set default to the literal
string "AUTHOR TEXT" as before.