Update atomics section in README.chplenv
Simplify the explanation for how we choose whether locks or intrinsics are
going to be used to implement intrinsics.

We added a few *32 bit platforms this release so it was no longer just linux32
that didn't support atomics. Also gcc added support for 64 bit intrinsics on
32 bit platforms with 4.8.1 so we do support intrinsics on 32 bit machines
depending on the target compiler.