Move all CommDiagnostics info into chpldocs and remove the technote.
doc/release/technotes/README.comm-diagnostics contained some information
(notably example programs and output) not in the CommDiagnostics.chpl
chpldocing.  Move that into CommDiagnostics chpldocs, and delete the
technote.  While here, I also did some rephrasing in the first paragraph
of the module-level chpldoc entry.

In addition, I moved the start of the module-level chpldoc to the next
line below the comment opener, and reduced the indentation throughout
that entry by 1 space.  This seemed to resolve a problem I was having
with the first line of the documentation being taken as a synopsis,
though I was making other changes at the same time and I may be mistaken
in this.  In any case, this indentation change hides the content changes
I made.  To see those, you'll need a diff command with a -b option.