Squash 'inline' markers on chpldoc'd procedures
This has been a contentious discussion, but my current thinking
is that the benefit of knowing something is inline'd for the
chpldoc client is minimal and that the lack of symmetry it
creates for modules like Math.chpl is unfortunate.  Until such
time as we have a compelling example that wants its inlines
documented badly, I've proposed we squash them.

I'm not entirely certain what the inlineTest*'s do and whether
I've handled them 100% appropriately.  I just looked for
.good files under test/chpldoc/ with 'inline proc' in them,
edited, and they seemed to work.

Mostly, I wanted to see if this was the expected 5-minute fix (TM)
and still hope to generate 1.11's docs with it (as well as the
version of Assert.chpl that has its 'module Assert' declaration
and associated comments enabled).