Support more comment terminators
In chpldocing IO.chpl, I needed to describe the %/*/ format
string, which unfortunately has a nested comment start in it.
For the documentation for the release, I just put %/ * / and
then added a comment that the spaces were to be removed.

Here I modify the comment parsing to support 3 new types
of block comments besides /* */ :
 /** **/
 /*** ***/
 /**** ****/

All four of them can be nested as well, but (for example) /*
is not considered a comment start at all within a /** comment;
something like /** /** **/ **/ is comment nesting for /** **/ comments.

Also added a test case.

For this pull request to be simple, I havn't done 'make parser'
but I will once it is reviewed.